Moved DeathBeacon class to its own file
